Finished Connecting Front and Back-end

+++ buy-and-sell-backend/src/routes/deleteListing.js
@@ -1,3 +1,4 @@
+import * as admin from 'firebase-admin'
 import { db } from "../database";
 
 
@@ -5,10 +6,13 @@ export const deleteListingRoute = {
     method: 'DELETE',
     path: '/api/listings/{id}',
     handler: async (req, h) => {
+        const token = req.headers.authtoken;
+        const user = await admin.auth().verifyIdToken(token);
+        const userId = user.user_id;
         const { id } = req.params;
         await db.query(
-            'DELETE FROM listings WHERE id=?',
-            [id],
+            'DELETE FROM listings WHERE id=? AND user_id=?',
+            [id, userId],
         )
         return { message: 'Success!'}
     }
